feat: move subtitle.py to asmrmanager/common
